test(sampling): use valid seeds in narrow cache regression
Keep the cache-width regression focused on cache validation by supplying an int64 seed. Remove duplicated wording from the sampler design document.

+++ tests/e2e/pull_request/one_card/test_gumbel_sampling.py
@@ -607,7 +607,7 @@ class TestGumbelSampling:
                 logits,
                 mapping,
                 torch.ones(1, device=DEVICE),
-                mapping,
+                mapping.to(torch.int64),
                 mapping,
                 True,
                 logits_cache=torch.empty(1, 1, 31, device=DEVICE),
